refactor: spread out code a bit
